test(`flet build`): group iOS build tests in test_build_ios.py
Rename test_build_ios_output.py to test_build_ios.py and put the
existing cases in a TestBuildOutput class, matching the grouping style
used by test_project_dependencies.py. The file now covers iOS builds
generally, so later cases get a class of their own instead of a new
file, with make_command shared between them.
